Moved and merged some of the image tests
