--- CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/algorithms/images/FloodFillMat.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d6760e4d42df9d34d854f4185d6538d0
