--- CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/algorithms/images/ImageBlur.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1bfa863137e2d1b4cb2c02cd592e8432
