+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/algorithms/images/ImageResultPixels.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 77cbb1e4e41466249b51ce154e2d2216
