+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b9f4fad6e53adcd40bd5871ea79ee194
+guid: 0bade508efc73b34fa94fbcc58c177e3
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
