--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/CutoutGenerationTest.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 432d143c1853f6a498915bfcdbe312c7
