--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/FilterTest.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 47e9a2638f13d354599d79a4e6277adb
