--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/FloodFillVsFloodFillOnMat.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 56e5418756a868b408bf7e2a3e515bad
