--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/GuidedFilterTest.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e320ce68ae3db27449dd3e51c4f29a96
