--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/ImageAlphaMattingTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1cb7338e47392004eb2e8b6a0f374a3c
