--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/ImageCompareTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 13f10bdecdc1b9844b8be26c36f6a834
