--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/ImageUtilityTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 46cb31caac6de274bad32a18e3970d84
