--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/MyImageFile.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 27296beaabf55c64991778005a748d49
