--- CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/TrimapGenerationTest.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a363918a1bd11f14fb359a45b514f394
