+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9208009c9923a54429cb9849b0b2863a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
