+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/FilterTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 112a082a832f0b0458de7940c930a9d1
