+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/ImageAlphaMattingTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 177a0d8e9cf1b764995cbeb3bbf856f6
