+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/ImageCompareTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 54b6ad529eed6fd4c993a49bfb84878e
