+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/ImageUtilityTests.cs
@@ -3,7 +3,7 @@ using com.csutil.algorithms.images;
 using StbImageWriteSharp;
 using Xunit;
 
-namespace com.csutil.tests.AlgorithmTests {
+namespace com.csutil.tests.AlgorithmTests.images {
 
     public class ImageUtilityTests {
