+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/ImageUtilityTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6c31711202c41ce448d49b74bdccb977
