+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/MyImageFile.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8dd2b65324de98a479d1d9c4384191cf
