+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/algorithms/images/TrimapGenerationTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 77221545db76aca4eb8879f62f9ef4c1
